--- aks-node-controller/pkg/json.go
@@ -1,22 +1,0 @@
-package pkg
-
-import (
-	aksnodeconfigv1 "github.com/Azure/agentbaker/aks-node-controller/pkg/gen/aksnodeconfig/v1"
-	"google.golang.org/protobuf/encoding/protojson"
-)
-
-func Marshal(cfg *aksnodeconfigv1.Configuration) ([]byte, error) {
-	options := protojson.MarshalOptions{
-		UseEnumNumbers: false,
-		UseProtoNames:  true,
-		Indent:         "  ",
-	}
-	return options.Marshal(cfg)
-}
-
-func Unmarshal(data []byte) (*aksnodeconfigv1.Configuration, error) {
-	cfg := &aksnodeconfigv1.Configuration{}
-	options := protojson.UnmarshalOptions{}
-	err := options.Unmarshal(data, cfg)
-	return cfg, err
-}
